% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{plate.layout}
\alias{plate.layout}
\title{Generate a box/plate based layout}
\usage{
plate.layout(rows, cols)
}
\arguments{
\item{rows}{numeric; the number of rows}

\item{cols}{numeric; the number of columns}
}
\value{
a \code{data.frame} representing a box/plate based layout with both 'alphanumeric' and 'row_col' labels.
}
\description{
Generate a box/plate based layout for use with organizing sample storage (cryobox) or experimental culture plates (96-well).
}
\examples{
plate.layout(9,9)
plate.layout(8,12)

}
