% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{plate.plot}
\alias{plate.plot}
\title{Plot a box/plate based layout}
\usage{
plate.plot(plate, row.label = c("alpha", "numeric"))
}
\arguments{
\item{plate}{A \code{data.frame} as returned from \link{plate.layout}.}

\item{row.label}{Character string; row-labeling style:
\itemize{
\item 'alpha' (default) for letter-based row-labeling; 96-well plate style.
\item 'numeric' for number-based row-labeling
}}
}
\value{
a \link[ggplot2]{ggplot2} object
}
\description{
Plot a basic box/plate based layout; if metadata is added to a \link{plate.layout}, those additional columns can be used as additional plot aesthetics.
}
\examples{
##a standard 9x9 storage box
plate<-plate.layout(9,9)
plate

##skip every third spot
skip<-c(3,6,9)
for(i in seq((9-1))){
skip<-c(skip,c(3,6,9)+(i*9))
}

##'subject.id'
plate$subject.id<-NA
for(i in seq(skip)){
plate.index<-c(skip[i]-2,skip[i]-1)
plate$subject.id[plate.index]<-"Subject"
}
subject.ids<-paste0(na.omit(plate)[['subject.id']],na.omit(plate)[['row']])
plate[!is.na(plate$subject.id),'subject.id']<-subject.ids

##'visit.id'
plate$visit.id<-NA
for(i in unique(na.omit(plate$subject.id))){
visit.index<-grep(i,plate$subject.id)
plate[visit.index[1:2],'visit.id']<-"V1"
plate[visit.index[3:4],'visit.id']<-"V2"
plate[visit.index[5:6],'visit.id']<-"V3"
}

plate

##basic plot; empty box
plate.plot(plate,row.label = "n")

##plot with added geom using plate metadata
plate.plot(plate,row.label = "n") +
ggplot2::geom_point(
data=na.omit(plate),
mapping=ggplot2::aes(
col,
row,
color=subject.id,
shape=visit.id),
size=10)

}
