% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slap.R
\name{slap}
\alias{slap}
\alias{\%!\%}
\alias{\%!!\%}
\title{Slap Operator}
\usage{
expr \%!\% message

expr \%!!\% message
}
\arguments{
\item{expr}{An expression or quosure to evaluate carefully}

\item{message}{A message meant to be formatted by \code{\link[cli:cli_bullets]{cli::cli_bullets()}}

When \code{expr} generates an error, the \verb{\%!\%} and \verb{\%!!\%} operators
catch it and embed it in a new error thrown by \code{\link[cli:cli_abort]{cli::cli_abort()}}
with \code{message}.

When the current environment has an \code{error_call} object, it is
used as the \code{call} argument of \code{\link[cli:cli_abort]{cli::cli_abort()}}}
}
\description{
Slap Operator
}
\examples{

g <- function() {
  stop("ouch")
}
h <- function(error_call = rlang::caller_env()) {
  g() \%!\% "bam"
}
f <- function() {
  h()
}

\dontrun{
  f()
}

}
