% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slap.R
\name{\%!\%}
\alias{\%!\%}
\title{Alternative to withCallingHandlers}
\usage{
expr \%!\% message
}
\arguments{
\item{expr}{An \link[rlang:topic-defuse]{expression} or
\link[rlang:topic-quosure]{quosure} to evaluate.}

\item{message}{It is formatted via a call to \code{\link[cli:cli_bullets]{cli_bullets()}}.}

\item{frame}{The throwing context. Used as default for
\code{.trace_bottom}, and to determine the internal package to mention
in internal errors when \code{.internal} is \code{TRUE}.}
}
\description{
Alternative to withCallingHandlers
}
\examples{
g <- function() {
  stop("ouch")
}
f <- function(error_call = current_env()) {
  g() \%!\% "bam"
}
h <- function() {
  rlang::local_error_call(quote(foo))

  g() \%!\% "bam"
}

\dontrun{
  f()
  h()
}

}
