% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slap.R
\name{slap}
\alias{slap}
\alias{\%!\%}
\title{Slap Operator}
\usage{
expr \%!\% message
}
\arguments{
\item{expr}{An \link[rlang:topic-defuse]{expression} or
\link[rlang:topic-quosure]{quosure} to evaluate.}

\item{message}{It is formatted via a call to \code{\link[cli:cli_bullets]{cli_bullets()}}.}
}
\description{
Slap Operator
}
\examples{
g <- function() {
  stop("ouch")
}
f <- function(error_call = current_env()) {
  g() \%!\% "bam"
}
h <- function() {
  rlang::local_error_call(quote(foo()))

  g() \%!\% "bam"
}

\dontrun{
  f()
  h()
}

}
